+++ src/Hooks/Menu.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace AveryAbbott\WindowsDhcp\Hooks;
 
 use AveryAbbott\WindowsDhcp\Models\DhcpScope;